+++ src/test/java/se/kth/iv1350/pos/model/RegisterTest.java
@@ -15,7 +15,7 @@ public class RegisterTest {
 
     @Test
     public void testIncreaseAmount() {
-        float amount = 100;
+        double amount = 100;
         instance.increaseAmount(amount);
         assertEquals(100, instance.amount, "Amount was not increased");
     }
